[receiver/github] fix queue span lengths caused by GitHub event timestamp bug (#39972)
#### Description

GitHub events from skipped jobs sometimes have the createdAt timestamp
to be AFTER the startedAt timestamp. This causes the created span to
having a time of some crazy long time, while the duration attribute is
negative. This change accounts for that inconsistent behavior from
GitHub and adds the corresponding tests.


#### Link to tracking issue
Fixes #39824

#### Testing

Added the tests in the source code.

#### Documentation

Minor comment around the codeblock for context of why.