[processor/datadogsemantics] Make defaults explicit and add test (#39596)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Explicitly set defaults for each field - whatever the Datadog backend
would populate if you send the field as blank; e.g., if we send a blank
env, it would show up in the UI as the string `default`, so we should
surface that to the user with this processor. Also, populate `datadog.*`
fields even if they would be blank (this includes overriding them even
if the result would be blank if the `overrideIncomingDatadogFields`
option is set, for consistency)

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Co-authored-by: Pablo Baeyens <pbaeyens31+github@gmail.com>